fix exception for deserializing range facets with big numbers

+++ commercetools.Sdk/commercetools.Sdk.Domain/Query/FilteredFacetResult.cs
@@ -3,7 +3,7 @@
     [TypeMarker("filter")]
     public class FilteredFacetResult : FacetResult
     {
-        public int Count { get; set; }
-        public int ProductCount { get; set; }
+        public long Count { get; set; }
+        public long ProductCount { get; set; }
     }
 }
