+++ commercetools.Sdk/commercetools.Sdk.Domain/ShippingMethods/CartValueShippingRatePriceTier.cs
@@ -3,7 +3,7 @@
     [TypeMarker("CartValue")]
     public class CartValueShippingRatePriceTier : ShippingRatePriceTier
     {
-        public int MinimumCentAmount { get; set; }
+        public long MinimumCentAmount { get; set; }
         public Money Price { get; set; }
         public bool IsMatching { get; set; }
     }
